#164 Finish implement CreateGraphingDataPoints which return array of JSX DataPoints, update some docstrings.
