+++ my-webxr-app/src/App.tsx
@@ -2,16 +2,18 @@ import { Sky } from '@react-three/drei';
 import { Canvas } from '@react-three/fiber';
 import { Controllers, VRButton, XR } from '@react-three/xr';
 import { openDB } from 'idb';
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { LocalCsvReader, UrlCsvReader } from './components/CsvReader';
-import GraphingDataPoint from './components/GraphingDataPoint';
 import GraphingDataPointMenu from './components/GraphingDataPointMenu';
 import Floor from './components/Floor';
 import GenerateXYZ from './components/GenerateXYZ';
-import createPosition from './components/Positions';
 import { PointSelectionProvider } from './contexts/PointSelectionContext';
 import './styles.css';
 import TestingOptions from './testing/TestingOptions';
+import {
+  createGraphingDataPoints,
+} from './utils/CreateGraphingDataPoints';
+import DataPoint from './repository/DataPoint';
 
 // minNum and maxNum will be from the csv file, just hardcoded for now
 const minNum: number = -10;
@@ -37,45 +39,6 @@ export default function App() {
   const dbName = 'CsvDataBase';
   const storeName = 'CsvData';
 
-  // hard coded data. example data would be replaced with PCA results for coordinates
-  // also would like to make a new type for GenerateXYZ info, allowing for easier use
-  const exampleData = [[-1, -1, -1], [2, 3, 0], [4, 3, 0], [1, 1, 1], [3, 2, 2]];
-  const datapoint1 = createPosition({
-    data: exampleData[0],
-    AxisStartPoints: [startPointX, startPointY, startPointZ],
-    length: Length,
-    scale: scaleFactor,
-    max: maxNum,
-  });
-  const datapoint2 = createPosition({
-    data: exampleData[1],
-    AxisStartPoints: [startPointX, startPointY, startPointZ],
-    length: Length,
-    scale: scaleFactor,
-    max: maxNum,
-  });
-  const datapoint3 = createPosition({
-    data: exampleData[2],
-    AxisStartPoints: [startPointX, startPointY, startPointZ],
-    length: Length,
-    scale: scaleFactor,
-    max: maxNum,
-  });
-  const datapoint4 = createPosition({
-    data: exampleData[3],
-    AxisStartPoints: [startPointX, startPointY, startPointZ],
-    length: Length,
-    scale: scaleFactor,
-    max: maxNum,
-  });
-  const datapoint5 = createPosition({
-    data: exampleData[4],
-    AxisStartPoints: [startPointX, startPointY, startPointZ],
-    length: Length,
-    scale: scaleFactor,
-    max: maxNum,
-  });
-
   // Initialize the database and store for csv data
   useEffect(() => {
     const initializeDB = async () => {
@@ -90,6 +53,27 @@ export default function App() {
     };
     initializeDB();
   }, [dbName, storeName]);
+
+  // TEST PlottedData
+  const exampleDataPoints = [[-1, -1, -1], [2, 3, 0], [4, 3, 0], [1, 1, 1], [3, 2, 2]];
+  const dataPoints = exampleDataPoints.map((point) => new DataPoint(point[0], point[1], point[2]));
+
+  const [plottedDataPoints, setPlottedDataPoints] = useState<JSX.Element[]>([]);
+  useEffect(() => {
+    if (dataPoints.length > 0) {
+      const plottedPoints = createGraphingDataPoints(
+        dataPoints,
+        'columnX',
+        'columnY',
+        'columnZ',
+        [startPointX, startPointY, startPointZ],
+        Length,
+        scaleFactor,
+        maxNum,
+      );
+      setPlottedDataPoints(plottedPoints);
+    }
+  }, [dataPoints]);
   return (
     <>
       <div>
@@ -118,7 +102,7 @@ export default function App() {
             <ambientLight />
             <pointLight position={[10, 10, 10]} />
             <Controllers />
-
+            {plottedDataPoints}
             <GenerateXYZ
               minValue={minNum}
               maxValue={maxNum}
@@ -129,59 +113,6 @@ export default function App() {
               endPoint={Length}
               radius={radius}
               labelOffset={labelOffset}
-            />
-
-            {/* Temporary display/test of the data points.
-              These will eventually be created by the plot itself */}
-            <GraphingDataPoint
-              id={0}
-              marker="circle"
-              color="gray"
-              columnX="John Doe"
-              columnY="cmpt 145"
-              columnZ="97"
-              meshProps={{ position: datapoint1 }}
-              actualData={exampleData[0]}
-            />
-            <GraphingDataPoint
-              id={1}
-              marker="circle"
-              color="gray"
-              columnX="Bob Johnson"
-              columnY="math 110"
-              columnZ="81"
-              meshProps={{ position: datapoint2 }}
-              actualData={exampleData[1]}
-            />
-            <GraphingDataPoint
-              id={2}
-              marker="circle"
-              color="gray"
-              columnX="Bob John"
-              columnY="math 116"
-              columnZ="87"
-              meshProps={{ position: datapoint3 }}
-              actualData={exampleData[2]}
-            />
-            <GraphingDataPoint
-              id={3}
-              marker="circle"
-              color="gray"
-              columnX="Alice Smith"
-              columnY="stat 245"
-              columnZ="75"
-              meshProps={{ position: datapoint4 }}
-              actualData={exampleData[3]}
-            />
-            <GraphingDataPoint
-              id={4}
-              marker="circle"
-              color="gray"
-              columnX="Bob Smith"
-              columnY="math 115"
-              columnZ="85"
-              meshProps={{ position: datapoint5 }}
-              actualData={exampleData[4]}
             />
 
             <GraphingDataPointMenu position={[0, 2.2, -0.75]} />
