+++ my-webxr-app/src/data/DataAbstractor.tsx
@@ -9,8 +9,6 @@ export default interface DataAbstractor {
   getAvailableFields: () => Promise<string[]>;
   storeCSV: (batchItems: Array<Array<string> | Array<number>>) => Promise<boolean>;
   storePCA: (columnNames: string[]) => Promise<boolean>;
-  selectAxes: (columnXName: string, columnYName: string, columnZName: string) => Promise<boolean>;
-  selectPCA: (columnXName: string, columnYName: string, columnZName: string) => Promise<boolean>;
   calculateStatistics:() => Promise<boolean>;
   createDataPointsFrom3Columns:(
     columnX: string,
